requirements.txt
aiogram
python-dotenv
asyncpg
pytest
pytest-asyncio